image diffusion design change
