+++ src/assets/css/styles.css
@@ -1118,7 +1118,7 @@ textarea {
 }
 
 .prem-img-promptbox textarea {
-  @apply py-2 px-4 text-sm text-ghostwhite rounded-lg custom-scroll resize-none;
+  @apply py-2 px-4 text-sm text-ghostwhite custom-scroll resize-none;
 }
 .prem-audio-box {
   @apply pt-6 pb-10 px-8 rounded-bl rounded-br rounded-tr custom-scroll break-words;
